requirements.txt
#python-dotenv
#openai
#langchain
#langchain-openai
#langchain-chroma
#langgraph
#langserve[all]
#fastapi
#uvicorn
#tiktoken
#pdfplumber
#unstructured
#python-docx
#openpyxl

python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
langserve==0.3.1
langgraph==0.2.32
openai==1.43.0
chromadb==0.5.5
tiktoken==0.7.0
pydantic==2.9.2
